pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = "Gerenciador de livros MADR. Projeto de conclusão do curso FastAPI do Zero do dunossauro"
authors = [
    {name = "Thiago Nebuloni",email = "thiago.nebuloni@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.1"
pytest-asyncio = "^1.0.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scoop = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre-format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
concurrency = ['thread', 'greenlet']
